Move stuff around, not perfect yet..
