+++ document/src/test/java/com/yahoo/document/json/JsonReaderTestCase.java
@@ -30,6 +30,8 @@ import com.yahoo.document.datatypes.StringFieldValue;
 import com.yahoo.document.datatypes.Struct;
 import com.yahoo.document.datatypes.TensorFieldValue;
 import com.yahoo.document.datatypes.WeightedSet;
+import com.yahoo.document.json.readers.DocumentParseInfo;
+import com.yahoo.document.json.readers.SingleValueReader;
 import com.yahoo.document.update.AddValueUpdate;
 import com.yahoo.document.update.ArithmeticValueUpdate;
 import com.yahoo.document.update.ArithmeticValueUpdate.Operator;
@@ -63,6 +65,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import static com.yahoo.document.json.readers.SingleValueReader.*;
 import static com.yahoo.test.json.JsonTestHelper.inputJson;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
@@ -209,7 +212,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -225,7 +228,7 @@ public class JsonReaderTestCase {
                         + "\"put\": \"id:unittest:smoke::whee\""
                         + "}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -239,7 +242,7 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:smoke::whee\","
                         + " \"fields\": {}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -255,7 +258,7 @@ public class JsonReaderTestCase {
                         + "\"sandra\": \"person\","
                         + " \"cloud\": \"another person\"}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -275,7 +278,7 @@ public class JsonReaderTestCase {
                         + " \"person\","
                         + " \"another person\"]}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
         r.readUpdate(parseInfo.fieldsBuffer, doc);
@@ -291,7 +294,7 @@ public class JsonReaderTestCase {
                         + " \"person\": 37,"
                         + " \"another person\": 41}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
         r.readUpdate(parseInfo.fieldsBuffer, doc);
@@ -320,7 +323,7 @@ public class JsonReaderTestCase {
                         + " \"element\": \"person\","
                         + " \"increment\": 13}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
@@ -346,13 +349,13 @@ public class JsonReaderTestCase {
     @Test
     public final void testArithmeticOperators() {
         Tuple2[] operations = new Tuple2[] {
-                new Tuple2<String, Operator>(JsonReader.UPDATE_DECREMENT,
+                new Tuple2<String, Operator>(UPDATE_DECREMENT,
                         ArithmeticValueUpdate.Operator.SUB),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_DIVIDE,
+                new Tuple2<String, Operator>(UPDATE_DIVIDE,
                         ArithmeticValueUpdate.Operator.DIV),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_INCREMENT,
+                new Tuple2<String, Operator>(UPDATE_INCREMENT,
                         ArithmeticValueUpdate.Operator.ADD),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_MULTIPLY,
+                new Tuple2<String, Operator>(UPDATE_MULTIPLY,
                         ArithmeticValueUpdate.Operator.MUL) };
         for (Tuple2<String, Operator> operator : operations) {
             InputStream rawDoc = new ByteArrayInputStream(
@@ -362,7 +365,7 @@ public class JsonReaderTestCase {
                             + " \"" + (String) operator.first + "\": 13}}}}"));
 
             JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-            JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+            DocumentParseInfo parseInfo = r.parseDocument().get();
             DocumentType docType = r.readDocumentType(parseInfo.documentId);
             DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
@@ -396,7 +399,7 @@ public class JsonReaderTestCase {
                         + " \"element\": 3,"
                         + " \"assign\": \"nalle\"}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
@@ -434,7 +437,7 @@ public class JsonReaderTestCase {
                         + " \"nalle\": 2,"
                         + " \"tralle\": 7 }}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -455,7 +458,7 @@ public class JsonReaderTestCase {
                         + " \"nalle\","
                         + " \"tralle\"]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -476,7 +479,7 @@ public class JsonReaderTestCase {
                         + " { \"key\": \"nalle\", \"value\": \"kalle\"},"
                         + " { \"key\": \"tralle\", \"value\": \"skalle\"} ]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -495,7 +498,7 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:testsinglepos::bamf\","
                         + " \"fields\": { \"singlepos\": \"N63.429722;E10.393333\" }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -512,7 +515,7 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:testsinglepos::bamf\","
                         + " \"fields\": { \"singlepos\": \"W46.63;S23.55\" }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -533,7 +536,7 @@ public class JsonReaderTestCase {
                         + "\""
                         + " }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -553,7 +556,7 @@ public class JsonReaderTestCase {
                         + "{ \"key\": \"bamse\", \"value\": [1, 2, 3] }"
                         + "]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
@@ -575,7 +578,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": {"
                         + " \"assign\": \"orOther\" }}" + " }"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
         r.readUpdate(parseInfo.fieldsBuffer, doc);
@@ -594,7 +597,7 @@ public class JsonReaderTestCase {
                         + "{ \"key\": \"bamse\", \"value\": [1, 2, 3] }"
                         + "]}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
         r.readUpdate(parseInfo.fieldsBuffer, doc);
@@ -618,7 +621,7 @@ public class JsonReaderTestCase {
                         + " \"person\": 37,"
                         + " \"another person\": 41}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
         r.readUpdate(parseInfo.fieldsBuffer, doc);
@@ -819,7 +822,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"smething\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         exception.expect(NullPointerException.class);
@@ -848,7 +851,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         r.readPut(parseInfo.fieldsBuffer, put);
